finish clang-tidy transformations

+++ src/special_functions/test/tstF1.cc
@@ -39,7 +39,7 @@ void tstF1(UnitTest &ut) {
 
   ofstream out("debug.dat");
   for (size_t j = 0; j < 300; ++j) {
-    double eta = -10.0 + 0.1 * j;
+    double eta = -10.0 + 0.1 * static_cast<double>(j);
     out << eta << ' ' << F1(eta) << endl;
   }
 }
