+++ src/special_functions/test/tstF12.cc
@@ -40,7 +40,7 @@ void tstF12(UnitTest &ut) {
 
   ofstream out("debug.dat");
   for (size_t j = 0; j < 300; ++j) {
-    double eta = -10.0 + 0.1 * j;
+    double eta = -10.0 + 0.1 * static_cast<double>(j);
     out << eta << ' ' << F12(eta) << endl;
   }
 }
