+++ src/special_functions/test/tstF32.cc
@@ -39,13 +39,12 @@ void tstF32(UnitTest &ut) {
 
   ofstream out("debug.dat");
   for (size_t j = 0; j < 300; ++j) {
-    double eta = -10.0 + 0.1 * j;
+    double eta = -10.0 + 0.1 * static_cast<double>(j);
     out << eta << ' ' << F32(eta) << endl;
   }
 }
 
 //------------------------------------------------------------------------------------------------//
-
 int main(int argc, char *argv[]) {
   ScalarUnitTest ut(argc, argv, release);
   try {
